requirements.txt
streamlit==1.29.0
pandas==2.0.3
scikit-learn==1.3.0
xgboost==2.0.0
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.0